+++ src/dbnode/persist/fs/commitlog/reader.go
@@ -24,7 +24,6 @@ import (
 	"context"
 	"encoding/binary"
 	"errors"
-	"fmt"
 	"io"
 	"os"
 	"sync"
@@ -174,7 +173,6 @@ func (r *reader) Open(filePath string) (time.Time, time.Duration, int64, error) 
 	r.chunkReader.reset(fd)
 	info, err := r.readInfo()
 	if err != nil {
-		fmt.Println("read info error :(")
 		r.Close()
 		return timeZero, 0, 0, err
 	}
@@ -390,9 +388,6 @@ func (r *reader) handleDecoderLoopIterationEnd(arg decoderArg, outBuf chan<- rea
 	if arg.bytes != nil {
 		arg.bufPool <- arg.bytes
 	}
-	if arg.err != nil {
-		fmt.Println("decoder loop: ", err)
-	}
 	if outBuf != nil {
 		response.resultErr = err
 		outBuf <- response
@@ -495,10 +490,8 @@ func (r *reader) readChunk(buf []byte) ([]byte, error) {
 func (r *reader) readInfo() (schema.LogInfo, error) {
 	data, err := r.readChunk([]byte{})
 	if err != nil {
-		fmt.Println("FAIL")
 		return emptyLogInfo, err
 	}
-	fmt.Println("SUCCESS")
 	r.infoDecoderStream.Reset(data)
 	r.infoDecoder.Reset(r.infoDecoderStream)
 	logInfo, err := r.infoDecoder.DecodeLogInfo()
